test(ui): control the simulated terminal
Share one decorated-terminal fixture that resets NO_COLOR, TERM and FORCE_COLOR so headless caller settings cannot make colour assertions environment-dependent.
